ditched TaskPresenter, back to TaskAdapter
but this time, it truly is just a recyclerview adapter. no nonsensical
nested asynctask, no superfluous list manipulation methods. it implements
just the abstract methods and nothing else. but it does have a reference
to the presenter in order to make the delete buttons functional. i think
that's unavoidable.

i think i'm ready to proceed to the next step in the codelab

- ActivityModule: moved the context to the global module
- GcmNmApp: global module now needs to be instantiated because it's not
  empty anymore
- MainContract.Scope: added a provider for a list shared by the
  recyclerview adapter and the view. the view is responsible for all
  the content management. the adapter only reads from it.
- MainPresenter: added deleteTask method
- MainState: added delete method that's only used if there's a loaded
  list waiting to be displayed by a recyclerview
- MainView: removed @PerActivity annotation because it's not actually
  provided by a @PerActivity component.
- MainView: all list manipulations now happen here
- MainView: made method signatures and nomenclature consistent
- MainView: added delete method
- TaskStorage: changed lifetime to @Singleton