came up with an alternate contract and implementation.
so i was flipping my phone and i realized that i always read from the
disk on config change. i thought about it and realized that there really
shouldn't be a separate view state and presenter state. so now the
presenter and the adapter share the List<T> with the adapter doing only
reads and the presenter doing all the adds and removes. i rewrote the
contract along the way, giving more descriptive interface names.

the view has become really really simple. i refactored the signature
to make them less similar (there's a lot of duplication in the signatures
of the presenter and model in the previous contract i wrote). the model
is gone. it is now clearer what the presenter really is responsible for:
input handling and backend-ui synchronization. i think i'm going to go
with this scheme from hereon.

i'm officially done with this project.

- manifest: added new activity entry
- MainActivity: added button listener directly from xml. i decided to
  break the consistency in this case because this really shouldn't be
  part of the contracts and i will be forced to inject a context into
  the presenter (which i want to avoid) if i include activity launching
  in the contract.
- layout_main: added button, tweaked other buttons' dimensions
- ActivityInjector: added AltWiring.Injector provider
- the rest is all-new stuff